Add Week04 pointer and variable demos, update C examples

+++ Week04/01-local-variables.c
@@ -26,14 +26,14 @@ void main(void) {
    unsigned char varchr6='g';
    unsigned char varchr7='h';
    printf(HEAD1);
-   printf("varchr0  =  %#4X = %c   %#16.16lX\n", varchr0, varchr0, &varchr0);
-   printf("varchr1  =  %#4X = %c   %#16.16lX\n", varchr1, varchr1, &varchr1);
-   printf("varchr2  =  %#4X = %c   %#16.16lX\n", varchr2, varchr2, &varchr2);
-   printf("varchr3  =  %#4X = %c   %#16.16lX\n", varchr3, varchr3, &varchr3);
-   printf("varchr4  =  %#4X = %c   %#16.16lX\n", varchr4, varchr4, &varchr4);
-   printf("varchr5  =  %#4X = %c   %#16.16lX\n", varchr5, varchr5, &varchr5);
-   printf("varchr6  =  %#4X = %c   %#16.16lX\n", varchr6, varchr6, &varchr6);
-   printf("varchr7  =  %#4X = %c   %#16.16lX\n", varchr7, varchr7, &varchr7);
+   printf("varchr0  =  %#4X = %c   %#16.16lX\n", varchr0, varchr0, (unsigned long)&varchr0);
+   printf("varchr1  =  %#4X = %c   %#16.16lX\n", varchr1, varchr1, (unsigned long)&varchr1);
+   printf("varchr2  =  %#4X = %c   %#16.16lX\n", varchr2, varchr2, (unsigned long)&varchr2);
+   printf("varchr3  =  %#4X = %c   %#16.16lX\n", varchr3, varchr3, (unsigned long)&varchr3);
+   printf("varchr4  =  %#4X = %c   %#16.16lX\n", varchr4, varchr4, (unsigned long)&varchr4);
+   printf("varchr5  =  %#4X = %c   %#16.16lX\n", varchr5, varchr5, (unsigned long)&varchr5);
+   printf("varchr6  =  %#4X = %c   %#16.16lX\n", varchr6, varchr6, (unsigned long)&varchr6);
+   printf("varchr7  =  %#4X = %c   %#16.16lX\n", varchr7, varchr7, (unsigned long)&varchr7);
 }
 
 /*
